DecalGeometry: Moved to examples
