--- docs/api/geometries/DecalGeometry.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<base href="../../" />
-		<script src="list.js"></script>
-		<script src="page.js"></script>
-		<link type="text/css" rel="stylesheet" href="page.css" />
-	</head>
-	<body>
-		[page:BufferGeometry] &rarr;
-
-		<h1>[name]</h1>
-
-		<div class="desc"> You can use this geometry to create a decal mesh, that serves different kinds of purposes e.g.:
-			adding unique details to models, performing dynamic visual environmental changes or covering seams. </div>
-
-		<h2>Example</h2>
-
-		[example:webgl_decals decals]
-
-		<code>var geometry = new THREE.DecalGeometry( mesh, position, orientation, size );
-		var material = new THREE.MeshBasicMaterial( { color: 0x00ff00 } );
-		var decal = new THREE.Mesh( geometry, material );
-		scene.add( decal );
-		</code>
-
-		<h2>Constructor</h2>
-
-		<h3>[name]( [page:Mesh mesh], [page:Vector3 position], [page:Euler orientation], [page:Vector3 size] )</h3>
-		<div>
-		mesh — Any mesh object.<br />
-		position — Position of the decal projector.<br />
-		orientation — Orientation of the decal projector.<br />
-		size — Size of the decal projector.<br />
-		</div>
-
-		<h2>Source</h2>
-
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
-	</body>
-</html>
